adjust

+++ ckb/celltype/types.go
@@ -308,9 +308,9 @@ func (p ProposeWitnessSliceDataObjectLL) ToMoleculeProposalCellData(incomeLockSc
 }
 
 type CalAccountCellExpiredAtParam struct {
-	Quote          uint32 `json:"quote"`
-	AccountCellCap uint32 `json:"account_cell_cap"`
-	PriceConfigNew uint32 `json:"price_config_new"`
+	Quote          uint64 `json:"quote"`
+	AccountCellCap uint64 `json:"account_cell_cap"`
+	PriceConfigNew uint64 `json:"price_config_new"`
 	// AccountBytesLen    uint32 `json:"account_bytes_len"`
-	PreAccountCellCap uint32 `json:"pre_account_cell_cap"`
+	PreAccountCellCap uint64 `json:"pre_account_cell_cap"`
 }
